dpif-netdev: use flow_table as indirect table
It is not memory efficient to store pointers in the distributor.
In this commit, we store a 2 Byte index which is the index into
flow_table. If the flow table is larger than 2^16, the rules
store in the high index entry will not be cached in distributor.

In cmap, we add two APIs to support find flow by index and find
index by flow. Since flow table is a cuckoo hash, it is possible
that keys are moved around and also keys to be rehashed.
In such case, distributor will have misses and
refresh. However, this should not happen frequently and
distributor as a cache should not cause functional error.

Signed-off-by: Yipeng Wang <yipeng1.wang@intel.com>